checkpoint XarExportContext.getRelativizedLSIDs() NOTE: Nobody ever constructs a XarExportContext AFAIK, so that's a problem.
