+++ api/src/org/labkey/api/exp/XarExportContext.java
@@ -3,6 +3,7 @@ package org.labkey.api.exp;
 import org.labkey.api.admin.FolderExportContext;
 import org.labkey.api.admin.LoggerGetter;
 import org.labkey.api.data.Container;
+import org.labkey.api.exp.xar.LSIDRelativizer;
 import org.labkey.api.security.User;
 
 import java.util.HashMap;
@@ -18,6 +19,7 @@ public class XarExportContext extends FolderExportContext
     private Set<Integer> _includedAssayRuns = new HashSet<>();
     private Map<Integer, Set<Integer>> _includedSamples = new HashMap<>();
     private Map<Integer, Set<Integer>> _includedDataClasses = new HashMap<>();
+    private final LSIDRelativizer.RelativizedLSIDs _relativizedLSIDs = new LSIDRelativizer.RelativizedLSIDs(LSIDRelativizer.FOLDER_RELATIVE);
 
     public XarExportContext(User user, Container c, Set<String> dataTypes, String format, LoggerGetter logger)
     {
@@ -57,5 +59,10 @@ public class XarExportContext extends FolderExportContext
     {
         if (includedDataClasses != null)
             _includedDataClasses = includedDataClasses;
+    }
+
+    public LSIDRelativizer.RelativizedLSIDs getRelativizedLSIDs()
+    {
+        return _relativizedLSIDs;
     }
 }
